Move DOM realated functions
