+++ js/app.js
@@ -1,4 +1,6 @@
-import {createElement as h, render, Component} from './component.js';
+import * as Protos from './prototypes.js';
+import { Component } from './component.js';
+import { createElement as h, render } from './miniDOM.js';
 
 const Title = (props) => h('h2', props, props.children);
 
@@ -14,7 +16,7 @@ class Counter extends Component {
 
     render() {
         return h('div', null,
-            h('div', null, `${this.state.value}`),
+            h('div', null, '{{ state.value }}'.interpolate(this)),
             h('button', {onClick: this.onPlusClick.bind(this)}, '+'),
         );
     }
