+++ js/component.js
@@ -1,45 +1,24 @@
-import {isStateLessComponent, isClass} from './utils.js';
-import { handleClass, createDOM, renderReactElement } from './miniDOM.js';
-
-let vDOM;
-
-function reRender(element) {
-    const key = element.props.key;
-    const domEl = document.getElementById(key);
-    const parent = domEl.parentNode;
-
-    parent.removeChild(domEl)
-    vDOM = renderReactElement(element);
-    
-    createDOM(vDOM, parent);
-}
-
-export const createElement = (element, props = null, ...children) => {
-    if (isClass(element)) {
-        return handleClass(element, props, children);
-    }
-    if (isStateLessComponent(element)) {
-        return element({...props, children});
-    }
-    
-    return {type: element, props, children};
-};
-
-export const render = (element, domElement) => {
-    vDOM = createElement(element);
-    createDOM(vDOM, domElement);
-};
+import { reRender } from './miniDOM.js';
+import { compareObjects } from './utils.js';
 
 export class Component {
     constructor(props) {
-        this.props = {
+        this.state = null;
+        this.props = props ? {
             key: Math.random().toString(36).substr(2, 9),
             ...props
-        };
+        } : {};
+    }
+
+    shouldUpdate(nextProps, nextState) {
+        return !compareObjects(nextProps, this.props) || !compareObjects(nextState, this.state);
     }
 
     setState(state) {
-        this.state = Object.assign({}, this.state, state);
-        reRender(this);
+        const nextState = {...this.state, ...state};
+        if (this.shouldUpdate(this.props, nextState)) {
+            this.state = nextState;
+            reRender(this);
+        }
     }
 }
