+++ js/utils.js
@@ -6,3 +6,5 @@ export function isClass(func) {
     return typeof func === 'function'
         && /^class\s/.test(Function.prototype.toString.call(func));
 }
+
+export const compareObjects = (obj1, obj2) => JSON.stringify(obj1) === JSON.stringify(obj2);
